+++ examples/spa/src/pages/Counter.tsx
@@ -31,4 +31,4 @@ export const Counter = component(({ signal }) => {
             </div>
         </>
     );
-});
+}, { name: 'Counter' });